REST API updates and added schedule controller

+++ src/main/kotlin/coffee/store/entity/CoffeeComponent.kt
@@ -15,7 +15,7 @@ data class CoffeeComponent(
         @JoinColumn(name = "id_ингредиента")
         val ingredient: Ingredient = Ingredient(),
         @Column(name = "количество")
-        val quantity: Int = 0,
+        val quantity: Double = 0.0,
         @Column(name = "порядок_добавления")
         val addingOrder: Int = 0,
 )
